Adjust map and encounter to bounded layer
This keeps the UI in sync with the declared aspect ratio.

Also, because the OK button was falling out of the dialog box, I increased the size of the popup to 90% of the size rather than 80%. In truth, the whole dialog box system requires much more careful attention, but this is stable at least.