Replace transparent.png with constraint-based sizing
It is wasteful to store a large empty image in the project just for the sake of setting a size. Constraints can be used, as shown in thsi commit. One could also generate a large empty image via game.plat.graphics.createCanvas(...), but this would be wasteful of video memory if all you are doing is using it for spacing.