Use PlayN JSON parser
This parser works on both Java and HTML platforms. With this commit, both platforms are loading the cockatrice encounter from encounters/cockatrice.json.

test-assets became more trouble than it was worth since the assets content could not easily be fetched using PlayN abstractions. Previously it was using the classloader, but this wouldn't work in a cross-platform way.