Use strings to represent skills
Previously we had a Skill enum, but this had the problem that it assumes we know the skills at compile time. Hence, if someone working on narrative wants to play with what skills are available, they need to work in the source code, not just in the narrative. Now, the set of skills is determined based on what are used in the narrative instead, and the code can just deal with whatever skills are available using narrative.skills().