Move json parsing tests to core
This required adding a java target dependency in the testing scope so that the tests could create a headless platform, which provides the json parser. However, this is worth it, since the production files being tests are in _this_ module, not the java one where the tests used to be.