Reset encounter on end-of-round phase
This fixes a defect where, if the player chose the same region as the previous encounter, the results from the past encounter were shown. The source of that defect was that 'context.encounter' was being held onto. Now, GameContext has been embedded with some validation logic---namely, erasing the encounter value at the end of the round.