Reintroduce the Skill abstraction
Previously, when the skill concept was only represented as a String, there had to be ad hoc, distributed logic to ensure that the strings were normalized---for example, to handle "weapon use" vs. "Weapon Use". By bringing the Skill class back (but not as an enum, like before), we now have one class that is clearly responsible for this normalization.