Require five points for victory.
Debug keys (1,2) will add points to the numbered player. The previous DebugMode has been renamed, since it only works without a game context. The new DebugKeys class is for keyboard commands in the game.

Supporting in-game debug keys required making the config accessible outside MonsterGame; a read-only interface was added to make it clear that consumers should not try to actually change the configuration. (The config is made by the platform-specific code and passed to MonsterGame.)