Show introductory slides at the start of the game
An RFuture is returned from the slideshow that will complete when the
show is done, whether it's because of the last slide or cancellation.

When a slide has popup text, the 'next' button is disabled during this
popup to prevent accidentally double-clicking, which would move to the
next slide before the popup is read.