Add core gameplay interactions
The images and text are not all there, and the layout is terrible, but
you can click through both players' turns. There is also no check for
the end of the game, so you can do this literally forever.

We left a TODO item in the files because the notebook flopping (making
sure the right notebook is on top) currently is unclean and has a
defect after the second round. We (Brennan and Paul) haven't the
resources to fix that up just yet, but we wanted to get this out there
for other people to branch off of.