Use native text input fields in HTML (#51)
This replaces the default TriplePlay pop-up text entry method with a native HTML text input overlays. This required writing a custom Html TriplePlay platform (HtmlTpPlatform), since TriplePlay only provides Java (LWJGL), SWT, and iOS TPPlatform implementations--all of which are in a state of broken, incidentally, so don't use them.

The HtmlTpPlatform will produce an HTML text input widget and place it as an overlay on a TriplePlay Field. The current implementation has a little bit over overlap, where you can see the underlying Field, but this is a minor visual issue and doesn't seem to break any behaviors.

One complication of this approach is that the overlay is not actually "connected" to any TriplePlay Group or PlayN Layer. As a result, it is practically impossible to animate its position, such as when a screen comes into view. (It might technically be possible, by running a timer that monitors the locations of all parents, but I did not go down this route.) Hence, the technique that is used is to show the fields only when the transition to name selection screen completes, and to hide them as the exit transition starts---in part because the API made this fairly straightforward compared to the alternatives. Again, there is a little bit of visual artifacting, as the native fields disappear prior to the screen swipe, but I think this is still preferable to having to deal with a pop-up window for text entry.