Demonstrate flipping child layers and TP roots
A critical part of this is that the grouplayers containing children (including roots) must not be clipped. Clipping is not aware of quad batches, so it will clip away the flipped pages. We should probably add a check in the constructor that none of the layers are group layers with clipping enabled, to prevent problems.