Add encounter page
This contains a significant change to the logic of the notebook. Now, instead of all pages being added by their upper-left corners, the "left" pages are added by their upper-rights, so that the rotation effect is correct. Without this, the left pages came out backward, since they were rotated around the y-axis.