Manage root disposal to prevent artifacting
Previously, visual junk would accumute due to the creation of roots that were overlapping and not managed. Now, the roots are removed from the iface while the layer is not visible. This eliminates the artifacting on the book. It is still visible in the control buttons, but those are just for the proof-of-concept screen anyway.