Update editor to new format (#55)
* Convert editor to use more implicit, less buggy structure

* Change story points to inspiration points

* Replace refresh methods with implementation for machines that don't have the method (for some reason)

* Change ReactionEditPane to accept a Story.

* Update reaction editor to allow multiple stories

One of the unit tests had to be disabled because the editor is being brought out of sync with the rest of the project. This has been documented with the test. The narrative parser should be updated in the next commit.

* Parse json with multiple stories for encounter

* Make test json have array

* Add and implement StoryDeck

The StoryDeck keeps track of unread stories and chooses one at random, reverting to all stories if one has been read.

* Add .equals() to StoryDeck

* Remove unread stories in order

* Allow for building multiple stories

* Add a test for order of StoryDeck stories

* Randomize the initial order of the story deck

This ensures that if someone plays the game again, they are not guaranteed to see the same story if they choose the same reaction.

* Ensure all encounters are met on subsequent passes through story deck

* Ensure story deck has at least one story

Without this assumption, the chooseOne method breaks.