Streamline build script
A build happens on master as part of the pull request process, but we
don't actually want to deploy the project unless this is a non-PR
build.

Also the previous approach was to clone and rebuild the whole playn
library. This is not necessary since Maven can go get the SNAPSHOT
build directly, just like it does when building via Maven locally.

This fixes #70.