Clean up character creation completion
This includes some refactorings to clarify the meanings of names and
selectors, and it eliminates an NPE that came up whenever an item was
deselected. This fixes #69.

This also eliminates a DRY violation in the skill and region button
implementation by introducing the concept of an abstract audible
toggle button.