Include meta tag indicating mobile-friendliness
This tag came from reading about responsive design. The new tag makes it clear that our page is designed to be shown on mobile as well as desktop browsers. In my local tests, this did not seem to actually change any behavior: in the game, we're already scaling based on the size of the viewport we're given. This tag is a clue to search engines that we are ready to roll on different sizes and shapes of screens.